# tests/fixtures/openapi.snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "M+ Trust Factor API",
    "version": "0.1.0",
    "description": "Public character/comparison routes with SWR refresh, Battle.net OAuth IAM, and permission-protected admin routes (shared admin key retained as emergency fallback)."
  },
...
  "paths": {
...
    "/api/v1/characters/autocomplete": {
      "get": {
        "tags": [
          "characters"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 8
            },
            "in": "query",
            "name": "region",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 96
            },
            "in": "query",
            "name": "query",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 96
            },
            "in": "query",
            "name": "q",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "suggestions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "realmSlug": {
                            "type": "string"
                          },
                          "region": {
                            "type": "string"
                          },
                          "classSlug": {
                            "type": [
                              "null",
                              "string"
                            ]
                          },
                          "specSlug": {
                            "type": [
                              "null",
                              "string"
                            ]
                          },
                          "avatarUrl": {
                            "type": [
                              "null",
                              "string"
                            ]
                          },
                          "classIconUrl": {
                            "type": [
                              "null",
                              "string"
                            ]
                          },
                          "source": {
                            "type": "string"
                          },
                          "kind": {
                            "type": "string"
                          },
                          "realmName": {
                            "type": [
                              "null",
                              "string"
                            ]
                          },
                          "label": {
                            "type": [
                              "null",
                              "string"
                            ]
                          }
                        },
                        "required": [
                          "name",
                          "realmSlug",
                          "region",
                          "classSlug",
                          "specSlug",
                          "avatarUrl",
                          "classIconUrl"
                        ],
                        "additionalProperties": true
                      }
                    }
                  },
                  "required": [
                    "suggestions"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/resolve": {
      "post": {
        "tags": [
          "characters"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 48
                  },
                  "realmSlug": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64
                  },
                  "region": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 8
                  },
                  "forceRetry": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "name",
                  "realmSlug",
                  "region"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "characterId": {
                      "type": "string"
                    },
                    "refreshId": {
                      "type": "string"
                    },
                    "profilePath": {
                      "type": "string"
                    },
                    "retryAfterMs": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "202": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "characterId": {
                      "type": "string"
                    },
                    "refreshId": {
                      "type": "string"
                    },
                    "profilePath": {
                      "type": "string"
                    },
                    "retryAfterMs": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "characterId": {
                      "type": "string"
                    },
                    "refreshId": {
                      "type": "string"
                    },
                    "profilePath": {
                      "type": "string"
                    },
                    "retryAfterMs": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "characterId": {
                      "type": "string"
                    },
                    "refreshId": {
                      "type": "string"
                    },
                    "profilePath": {
                      "type": "string"
                    },
                    "retryAfterMs": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "characterId": {
                      "type": "string"
                    },
                    "refreshId": {
                      "type": "string"
                    },
                    "profilePath": {
                      "type": "string"
                    },
                    "retryAfterMs": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "characterId": {
                      "type": "string"
                    },
                    "refreshId": {
                      "type": "string"
                    },
                    "profilePath": {
                      "type": "string"
                    },
                    "retryAfterMs": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/characters/{region}/{realm}/{name}/refresh": {
      "post": {
        "tags": [
          "characters"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 8
            },
            "in": "path",
            "name": "region",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            },
            "in": "path",
            "name": "realm",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 48
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "characterId": {
                      "type": "string"
                    },
                    "refreshStatus": {
                      "type": "string"
                    },
                    "job": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "jobId": {
                              "type": "string"
                            },
                            "queue": {
                              "type": "string"
                            },
                            "status": {
                              "type": "string"
                            },
                            "dedupeKey": {
                              "type": [
                                "null",
                                "string"
                              ]
                            },
                            "createdAt": {
                              "type": "string"
                            },
                            "startedAt": {
                              "type": [
                                "null",
                                "string"
                              ]
                            },
                            "finishedAt": {
                              "type": [
                                "null",
                                "string"
                              ]
                            },
                            "errorMessage": {
                              "type": [
                                "null",
                                "string"
                              ]
                            }
                          },
                          "additionalProperties": true
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "cooldownSecondsRemaining": {
                      "type": "number"
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/{region}/{realm}/{name}/refresh-status": {
      "get": {
        "tags": [
          "characters"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 8
            },
            "in": "path",
            "name": "region",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            },
            "in": "path",
            "name": "realm",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 48
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "characterId": {
                      "type": "string"
                    },
                    "refreshStatus": {
                      "type": "string"
                    },
                    "job": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "jobId": {
                              "type": "string"
                            },
                            "queue": {
                              "type": "string"
                            },
                            "status": {
                              "type": "string"
                            },
                            "dedupeKey": {
                              "type": [
                                "null",
                                "string"
                              ]
                            },
                            "createdAt": {
                              "type": "string"
                            },
                            "startedAt": {
                              "type": [
                                "null",
                                "string"
                              ]
                            },
                            "finishedAt": {
                              "type": [
                                "null",
                                "string"
                              ]
                            },
                            "errorMessage": {
                              "type": [
                                "null",
                                "string"
                              ]
                            }
                          },
                          "additionalProperties": true
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "cooldownSecondsRemaining": {
                      "type": "number"
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ]
}